Create SumClin_Assign1
This is for our first assignment creating a neural network of a single node acting as a logistic regression model. 

project Description
Question Generate a 2-class, 2-dimensional synthetic dataset consisting of

2000 instances, with 1000 instances per class, drawn from a Gaussian distribu-
tion. Use labels 0 and 1 to define instances of class 1 and class 2, respectively.

A sample 2-class, 2-dimensional synthetic dataset is shown in Figure 1.
To generate the dataset, PyTorch’s torch.normal module is useful. Define a
neural network with one node using PyTorch’s torch.nn module and a sigmoid
activation function, effectively creating a logistic regression model. Use 70% of
the synthetic data to train the logistic regression model, employing the binary
cross-entropy loss function. Use the remaining 30% of the data as test data to
compute the accuracy or performance of the trained model.